Refactor RowStore
