+++ src/row_store/Projection.h
@@ -52,4 +52,37 @@ IntermediateTable<T> *projection(IntermediateTable<T> &table, std::vector<uint64
 
     return result;
 }
+
+/// Copies the content of the columns that are referred in projectionParameters to a new intermediateTable-Objects and
+/// returns its address projected
+/// \param table table data of the original table
+/// \param projectionParameters vector of column indices that are contained in the result
+template <typename T>
+IntermediateTable_AVX<T> *projection_AVX(IntermediateTable_AVX<T> &table, std::vector<uint64_t> &projectionParameters) {
+    // Check for valid projection parameters
+    if (!columnIndicesValid(projectionParameters, table.getTupleWidth())) {
+        throw std::invalid_argument("Error! Invalid column indices!");
+    }
+
+    // Create empty intermediate table
+    auto result = new IntermediateTable_AVX<T>(projectionParameters.size(), table.getRowCount());
+
+    // Iterate over given table tuples
+    for (int i = 0; i < table.count(); i++) {
+        // Create empty (temporary) tuple
+        auto tuple = (T *)malloc(projectionParameters.size() * sizeof(T));
+
+        // Fill tuple with data from every row of table
+        for (int j = 0; j < projectionParameters.size(); j++) {
+            tuple[j] = table[i][projectionParameters[j]];
+        }
+
+        // Add tuple to result table
+        result->addRow(tuple);
+        free(tuple);
+    }
+
+    return result;
+}
+
 }  // namespace RowStore
